fix: fixed test to be runnable
