+++ src/test/java/sk/leziemevpezinku/spring/annotation/UnitTest.java
@@ -1,13 +1,17 @@
 package sk.leziemevpezinku.spring.annotation;
 
 import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.TYPE)
+import static java.lang.annotation.ElementType.TYPE;
+
+@Target(TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Tag("UnitTest")
+@ExtendWith(MockitoExtension.class)
 public @interface UnitTest { }
