+++ src/test/java/sk/leziemevpezinku/spring/mapper/AppParamMapperTest.java
@@ -1,9 +1,8 @@
-package sk.leziemevpezinku.spring.api;
+package sk.leziemevpezinku.spring.mapper;
 
 import org.junit.jupiter.api.Test;
 import sk.leziemevpezinku.spring.annotation.UnitTest;
 import sk.leziemevpezinku.spring.api.dto.AppParamDTO;
-import sk.leziemevpezinku.spring.mapper.AppParamMapper;
 import sk.leziemevpezinku.spring.model.AppParam;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
